+++ src/DrawTools/Pencil.java
@@ -7,7 +7,7 @@ import Estimators.LossEstimator;
 import java.util.ArrayList;
 
 public class Pencil extends PixelPencil{
-    public Pencil(float[][] image, float[][] canvas, ArrayList<float[][]> features,
+    public Pencil(double[][] image, double[][] canvas, ArrayList<double[][]> features,
                   int maxIter, int minLoss, int area_step, int area_size,
                   ToolParams toolParams, LossEstimator mainLossEstimator, ColorAdder colorAdder) {
         super(canvas, features, maxIter, minLoss, area_step, area_size, toolParams, mainLossEstimator, colorAdder);
